+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/DataFactory.java
@@ -25,9 +25,6 @@ import com.liferay.asset.kernel.model.AssetVocabularyModel;
 import com.liferay.blogs.constants.BlogsPortletKeys;
 import com.liferay.blogs.model.BlogsEntry;
 import com.liferay.blogs.model.BlogsEntryModel;
-import com.liferay.blogs.model.BlogsStatsUserModel;
-import com.liferay.blogs.model.impl.BlogsEntryModelImpl;
-import com.liferay.blogs.model.impl.BlogsStatsUserModelImpl;
 import com.liferay.blogs.social.BlogsActivityKeys;
 import com.liferay.commerce.currency.model.CommerceCurrencyModel;
 import com.liferay.commerce.currency.model.impl.CommerceCurrencyModelImpl;
@@ -76,12 +73,6 @@ import com.liferay.fragment.model.impl.FragmentCollectionModelImpl;
 import com.liferay.fragment.model.impl.FragmentEntryLinkModelImpl;
 import com.liferay.fragment.model.impl.FragmentEntryModelImpl;
 import com.liferay.friendly.url.model.FriendlyURLEntryLocalization;
-import com.liferay.friendly.url.model.FriendlyURLEntryLocalizationModel;
-import com.liferay.friendly.url.model.FriendlyURLEntryMappingModel;
-import com.liferay.friendly.url.model.FriendlyURLEntryModel;
-import com.liferay.friendly.url.model.impl.FriendlyURLEntryLocalizationModelImpl;
-import com.liferay.friendly.url.model.impl.FriendlyURLEntryMappingModelImpl;
-import com.liferay.friendly.url.model.impl.FriendlyURLEntryModelImpl;
 import com.liferay.hello.world.web.internal.constants.HelloWorldPortletKeys;
 import com.liferay.journal.constants.JournalActivityKeys;
 import com.liferay.journal.constants.JournalArticleConstants;
@@ -150,8 +141,6 @@ import com.liferay.portal.kernel.model.Role;
 import com.liferay.portal.kernel.model.RoleModel;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.model.UserModel;
-import com.liferay.portal.kernel.model.UserNotificationDeliveryConstants;
-import com.liferay.portal.kernel.model.UserNotificationDeliveryModel;
 import com.liferay.portal.kernel.model.UserPersonalSite;
 import com.liferay.portal.kernel.model.VirtualHostModel;
 import com.liferay.portal.kernel.model.role.RoleConstants;
@@ -164,7 +153,6 @@ import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.FastDateFormatFactoryUtil;
 import com.liferay.portal.kernel.util.FriendlyURLNormalizerUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.ObjectValuePair;
 import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.util.ReleaseInfo;
@@ -186,7 +174,6 @@ import com.liferay.portal.model.impl.ReleaseModelImpl;
 import com.liferay.portal.model.impl.ResourcePermissionModelImpl;
 import com.liferay.portal.model.impl.RoleModelImpl;
 import com.liferay.portal.model.impl.UserModelImpl;
-import com.liferay.portal.model.impl.UserNotificationDeliveryModelImpl;
 import com.liferay.portal.model.impl.VirtualHostModelImpl;
 import com.liferay.portal.upgrade.PortalUpgradeProcess;
 import com.liferay.portal.util.PropsValues;
@@ -407,20 +394,12 @@ public class DataFactory extends BaseDDMDataFactory {
 		return allAssetVocabularyModels;
 	}
 
-	public long getBlogsEntryClassNameId() {
-		return getClassNameId(BlogsEntry.class);
-	}
-
 	public long getCounterNext() {
 		return counter.get();
 	}
 
 	public int getMaxAssetPublisherPageCount() {
 		return BenchmarksPropsValues.MAX_ASSETPUBLISHER_PAGE_COUNT;
-	}
-
-	public int getMaxBlogsEntryCommentCount() {
-		return BenchmarksPropsValues.MAX_BLOGS_ENTRY_COMMENT_COUNT;
 	}
 
 	public int getMaxCommerceProductCount() {
@@ -881,31 +860,6 @@ public class DataFactory extends BaseDDMDataFactory {
 				PortletConstants.DEFAULT_PREFERENCES));
 
 		return portletPreferencesModels;
-	}
-
-	public List<BlogsEntryModel> newBlogsEntryModels(long groupId) {
-		List<BlogsEntryModel> blogEntryModels = new ArrayList<>(
-			BenchmarksPropsValues.MAX_BLOGS_ENTRY_COUNT);
-
-		for (int i = 1; i <= BenchmarksPropsValues.MAX_BLOGS_ENTRY_COUNT; i++) {
-			blogEntryModels.add(newBlogsEntryModel(groupId, i));
-		}
-
-		return blogEntryModels;
-	}
-
-	public BlogsStatsUserModel newBlogsStatsUserModel(long groupId) {
-		BlogsStatsUserModel blogsStatsUserModel = new BlogsStatsUserModelImpl();
-
-		blogsStatsUserModel.setStatsUserId(counter.get());
-		blogsStatsUserModel.setGroupId(groupId);
-		blogsStatsUserModel.setCompanyId(COMPANY_ID);
-		blogsStatsUserModel.setUserId(SAMPLE_USER_ID);
-		blogsStatsUserModel.setEntryCount(
-			BenchmarksPropsValues.MAX_BLOGS_ENTRY_COUNT);
-		blogsStatsUserModel.setLastPostDate(new Date());
-
-		return blogsStatsUserModel;
 	}
 
 	public GroupModel newCommerceCatalogGroupModel(
@@ -1535,71 +1489,6 @@ public class DataFactory extends BaseDDMDataFactory {
 		fragmentEntryModel.setStatus(WorkflowConstants.STATUS_APPROVED);
 
 		return fragmentEntryModel;
-	}
-
-	public FriendlyURLEntryLocalizationModel
-		newFriendlyURLEntryLocalizationModel(
-			FriendlyURLEntryModel friendlyURLEntryModel,
-			BlogsEntryModel blogsEntryModel) {
-
-		FriendlyURLEntryLocalizationModel friendlyURLEntryLocalizationModel =
-			new FriendlyURLEntryLocalizationModelImpl();
-
-		friendlyURLEntryLocalizationModel.setFriendlyURLEntryLocalizationId(
-			counter.get());
-		friendlyURLEntryLocalizationModel.setFriendlyURLEntryId(
-			friendlyURLEntryModel.getFriendlyURLEntryId());
-		friendlyURLEntryLocalizationModel.setGroupId(
-			friendlyURLEntryModel.getGroupId());
-		friendlyURLEntryLocalizationModel.setCompanyId(
-			friendlyURLEntryModel.getCompanyId());
-		friendlyURLEntryLocalizationModel.setClassNameId(
-			friendlyURLEntryModel.getClassNameId());
-		friendlyURLEntryLocalizationModel.setClassPK(
-			friendlyURLEntryModel.getClassPK());
-		friendlyURLEntryLocalizationModel.setLanguageId(
-			LocaleUtil.toLanguageId(LocaleUtil.getSiteDefault()));
-		friendlyURLEntryLocalizationModel.setUrlTitle(
-			blogsEntryModel.getUrlTitle());
-
-		return friendlyURLEntryLocalizationModel;
-	}
-
-	public FriendlyURLEntryMappingModel newFriendlyURLEntryMapping(
-		FriendlyURLEntryModel friendlyURLEntryModel) {
-
-		FriendlyURLEntryMappingModel friendlyURLEntryMappingModel =
-			new FriendlyURLEntryMappingModelImpl();
-
-		friendlyURLEntryMappingModel.setFriendlyURLEntryMappingId(
-			counter.get());
-		friendlyURLEntryMappingModel.setClassNameId(
-			friendlyURLEntryModel.getClassNameId());
-		friendlyURLEntryMappingModel.setClassPK(
-			friendlyURLEntryModel.getClassPK());
-		friendlyURLEntryMappingModel.setFriendlyURLEntryId(
-			friendlyURLEntryModel.getFriendlyURLEntryId());
-
-		return friendlyURLEntryMappingModel;
-	}
-
-	public FriendlyURLEntryModel newFriendlyURLEntryModel(
-		BlogsEntryModel blogsEntryModel) {
-
-		FriendlyURLEntryModel friendlyURLEntryModel =
-			new FriendlyURLEntryModelImpl();
-
-		friendlyURLEntryModel.setDefaultLanguageId("en_US");
-		friendlyURLEntryModel.setUuid(SequentialUUID.generate());
-		friendlyURLEntryModel.setFriendlyURLEntryId(counter.get());
-		friendlyURLEntryModel.setGroupId(blogsEntryModel.getGroupId());
-		friendlyURLEntryModel.setCompanyId(COMPANY_ID);
-		friendlyURLEntryModel.setCreateDate(new Date());
-		friendlyURLEntryModel.setModifiedDate(new Date());
-		friendlyURLEntryModel.setClassNameId(getClassNameId(BlogsEntry.class));
-		friendlyURLEntryModel.setClassPK(blogsEntryModel.getEntryId());
-
-		return friendlyURLEntryModel;
 	}
 
 	public GroupModel newGlobalGroupModel() {
@@ -2500,24 +2389,6 @@ public class DataFactory extends BaseDDMDataFactory {
 		return userModels;
 	}
 
-	public UserNotificationDeliveryModel newUserNotificationDeliveryModel(
-		String portletId) {
-
-		UserNotificationDeliveryModel userNotificationDeliveryModel =
-			new UserNotificationDeliveryModelImpl();
-
-		userNotificationDeliveryModel.setUserNotificationDeliveryId(
-			counter.get());
-		userNotificationDeliveryModel.setCompanyId(COMPANY_ID);
-		userNotificationDeliveryModel.setUserId(SAMPLE_USER_ID);
-		userNotificationDeliveryModel.setPortletId(portletId);
-		userNotificationDeliveryModel.setDeliveryType(
-			UserNotificationDeliveryConstants.TYPE_WEBSITE);
-		userNotificationDeliveryModel.setDeliver(true);
-
-		return userNotificationDeliveryModel;
-	}
-
 	public GroupModel newUserPersonalSiteGroupModel() {
 		return newGroupModel(
 			_userPersonalSiteGroupId, getClassNameId(UserPersonalSite.class),
@@ -2807,29 +2678,6 @@ public class DataFactory extends BaseDDMDataFactory {
 		assetVocabularyModel.setLastPublishDate(new Date());
 
 		return assetVocabularyModel;
-	}
-
-	protected BlogsEntryModel newBlogsEntryModel(long groupId, int index) {
-		BlogsEntryModel blogsEntryModel = new BlogsEntryModelImpl();
-
-		blogsEntryModel.setUuid(SequentialUUID.generate());
-		blogsEntryModel.setEntryId(counter.get());
-		blogsEntryModel.setGroupId(groupId);
-		blogsEntryModel.setCompanyId(COMPANY_ID);
-		blogsEntryModel.setUserId(SAMPLE_USER_ID);
-		blogsEntryModel.setUserName(SAMPLE_USER_NAME);
-		blogsEntryModel.setCreateDate(new Date());
-		blogsEntryModel.setModifiedDate(new Date());
-		blogsEntryModel.setTitle("Test Blog " + index);
-		blogsEntryModel.setSubtitle("Subtitle of Test Blog " + index);
-		blogsEntryModel.setUrlTitle("testblog" + index);
-		blogsEntryModel.setContent("This is test blog " + index + ".");
-		blogsEntryModel.setDisplayDate(new Date());
-		blogsEntryModel.setLastPublishDate(new Date());
-		blogsEntryModel.setStatusByUserId(SAMPLE_USER_ID);
-		blogsEntryModel.setStatusDate(new Date());
-
-		return blogsEntryModel;
 	}
 
 	protected CPFriendlyURLEntryModel newCPFriendlyURLEntryModel(
