+++ modules/util/portal-tools-sample-sql-builder/src/main/java/com/liferay/portal/tools/sample/sql/builder/SampleSQLBuilder.java
@@ -249,6 +249,8 @@ public class SampleSQLBuilder {
 					FreeMarkerUtil.process(
 						BenchmarksPropsValues.SCRIPT,
 						HashMapBuilder.<String, Object>put(
+							"blogDataFactory", new BlogDataFactory()
+						).put(
 							"csvFileWriter", csvFileWriter
 						).put(
 							"dataFactory", new DataFactory()
